+++ src/index.mjs
@@ -55,79 +55,11 @@ Hooks.on(`closeActiveEffectConfig`, function(app, html) {
 });
 
 Hooks.on(`suite-5e.dialog.add`, async function(id, actor, options) {
-    switch (id) {
-        case `suite-5e.dialog`:
-            ui.notifications.info(`Custom Dialog '${id}' added options: ${options.join(`, `)}`);
-            break;
-        case `suite-5e.armor`:
-            const armors = actor.system.traits.armorProf.value;
-            options.forEach(x => armors.add(x));
-            await actor.update({ 'system.traits.armorProf.value': armors });
-            break;
-        case `suite-5e.weapon`:
-            const weapons = actor.system.traits.weaponProf.value;
-            options.forEach(x => weapons.add(x));
-            await actor.update({ 'system.traits.armorProf.value': weapons });
-            break;
-        case `suite-5e.tool`:
-            const tools = actor.system.tools;
-            options.forEach(x => tools[x] = { value: 1 });
-            await actor.update({ 'system.tools': tools });
-            break;
-        case `suite-5e.language`:
-            const languages = actor.system.traits.languages.value;
-            options.forEach(x => languages.add(x));
-            await actor.update({ 'system.traits.languages.value': languages });
-            break;
-        case `suite-5e.save`:
-            const saveUpdates = {};
-            options.forEach(x => saveUpdates[`system.abilities.${x}.proficient`] = 1);
-            await actor.update(saveUpdates);
-            break;
-        case `suite-5e.skill`:
-            const skillUpdates = {};
-            options.forEach(x => skillUpdates[`system.skills.${x}.value`] = 1);
-            await actor.update(skillUpdates);
-            break;
-    }
+    dialogs.onDialogAdd(id, actor, options);
 });
 
 Hooks.on(`suite-5e.dialog.rem`, async function(id, actor, options) {
-    switch (id) {
-        case `suite-5e.dialog`:
-            ui.notifications.info(`Custom Dialog '${id}' removed options: ${options.join(`, `)}`);
-            break;
-        case `suite-5e.armor`:
-            const armors = actor.system.traits.armorProf.value;
-            options.forEach(x => armors.delete(x));
-            await actor.update({ 'system.traits.armorProf.value': armors });
-            break;
-        case `suite-5e.weapon`:
-            const weapons = actor.system.traits.weaponProf.value;
-            options.forEach(x => weapons.delete(x));
-            await actor.update({ 'system.traits.armorProf.value': weapons });
-            break;
-        case `suite-5e.tool`:
-            const filteredTools = Object.keys(actor.system.tools).filter(key => !options.includes(key));
-            const tools = filteredTools.length > 0 ? filteredTools.reduce((obj, key) => { obj[key] = actor.system.tools[key]; return obj; }) : {};
-            await actor.update({ 'system.tools': tools });
-            break;
-        case `suite-5e.language`:
-            const languages = actor.system.traits.languages.value;
-            options.forEach(x => languages.delete(x));
-            await actor.update({ 'system.traits.languages.value': languages });
-            break;
-        case `suite-5e.save`:
-            const saveUpdates = {};
-            options.forEach(x => saveUpdates[`system.abilities.${x}.proficient`] = 0);
-            await actor.update(saveUpdates);
-            break;
-        case `suite-5e.skill`:
-            const skillUpdates = {};
-            options.forEach(x => skillUpdates[`system.skills.${x}.value`] = 0);
-            await actor.update(skillUpdates);
-            break;
-    }
+    dialogs.onDialogRem(id, actor, options);
 });
 
 $.fn.textWidth = function(text, font) {